pyproject.toml
[project]
name = "x-digest"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "httpx>=0.28.1",
    "openai>=2.21.0",
    "python-dotenv>=1.2.1",
    "tenacity>=9.1.4",
    "tweepy>=4.16.0",
    "twikit>=2.3.3",
]
